Выполнила задание Модуляция
