+++ js/form-ad.js
@@ -1,0 +1,1 @@
+// форма объявления
